extract shared worker utils to weight_transfer.py
Move get_vllm_model and get_weights_iterator to weight_transfer.py so
both FileSystemWeightUpdateWorker and SparseFileSystemWeightUpdateWorker
share the same implementation. No duplicate code between the two workers.
