+++ src/prime_rl/inference/vllm/worker/filesystem.py
@@ -1,9 +1,11 @@
+from pathlib import Path
 from typing import TYPE_CHECKING
 
-from torch.nn import Module
-from vllm.model_executor.model_loader import DefaultModelLoader, get_model_loader
-
-from prime_rl.inference.vllm.worker.weight_transfer import load_weights_checkpoint_layerwise
+from prime_rl.inference.vllm.worker.weight_transfer import (
+    get_vllm_model,
+    get_weights_iterator,
+    load_weights_checkpoint_layerwise,
+)
 
 # This is to get type hints for the Worker class but not actually extend it at runtime as this is required by vLLM worker extension
 if TYPE_CHECKING:
@@ -27,26 +29,10 @@ class FileSystemWeightUpdateWorker(Worker):
 
     def update_weights_from_path(self, weight_path: str) -> None:
         """Update weights from a specified path in shared filesystem containing a HF-compatible checkpoint."""
-        # Get vLLM model runner and model
-        # When enforce_eager=True, model isn't wrapped by torch.compile so no .runnable attr
-        model_runner = self.model_runner
-        if hasattr(model_runner.model, "runnable"):
-            model = model_runner.model.runnable
-        else:
-            model = model_runner.model
-        assert isinstance(model, Module)
-
-        # Get vLLM model loader
-        model_loader = get_model_loader(self.load_config)
-        assert isinstance(model_loader, DefaultModelLoader)
-        local_source = DefaultModelLoader.Source(
-            weight_path,
-            revision=None,  # TODO: Check that this is correct or if we should use the default (model_config.revision)
-            prefix="",
-            fall_back_to_pt=getattr(model, "fall_back_to_pt_during_load", True),
-            allow_patterns_overrides=getattr(model, "allow_patterns_overrides", None),
+        model = get_vllm_model(self.model_runner)
+        weights_iterator = get_weights_iterator(
+            model, Path(weight_path), self.load_config, self.model_runner.model_config
         )
-        weights_iterator = model_loader._get_weights_iterator(local_source)
         load_weights_checkpoint_layerwise(
             model,
             weights_iterator,
